refactor(studio): dedupe contact validator + extract grid helpers + a11y label
Resolves CodeRabbit nitpicks on #1613.

1. responsibility.ts: extract `validateContactFields(contact, required)`
   helper from the duplicated `Rule.custom` bodies on `primaryContact`
   and `solutionStep.contact`. Same teaching messages, single source
   of truth.

2. launcher-card.tsx: add an explicit `aria-label` on the role=button
   Card so screen readers announce intent ("Maak nieuw '<title>'
   document aan") instead of concatenating child text.

3. launcher-grid.tsx: move `groupByUiGroup` + `searchTemplates` into a
   sibling `launcher-grid-helpers.ts`. Mirrors the
   `filter-launcher-templates.ts` precedent — pure helpers stay
   importable from vitest without dragging Sanity UI / runtime CSS
   into the module graph.
