+++ packages/sanity-studio/src/tools/launcher/launcher-grid.tsx
@@ -3,46 +3,11 @@ import {type JSX, useMemo, useState} from 'react'
 
 import type {LauncherTemplate} from '../../templates/types'
 import {LauncherCard} from './launcher-card'
+import {groupByUiGroup, searchTemplates} from './launcher-grid-helpers'
 
 export interface LauncherGridProps {
   templates: ReadonlyArray<LauncherTemplate>
   onSelect: (template: LauncherTemplate) => void
-}
-
-/**
- * Group templates by `ui.group`, preserving the first-seen group order.
- * Pure helper — exported for unit tests.
- */
-export function groupByUiGroup(
-  templates: ReadonlyArray<LauncherTemplate>,
-): {group: string; templates: LauncherTemplate[]}[] {
-  const order: string[] = []
-  const map = new Map<string, LauncherTemplate[]>()
-  for (const template of templates) {
-    const key = template.ui.group
-    if (!map.has(key)) {
-      map.set(key, [])
-      order.push(key)
-    }
-    map.get(key)!.push(template)
-  }
-  return order.map((group) => ({group, templates: map.get(group)!}))
-}
-
-/**
- * Case-insensitive substring search across the title, description,
- * group label, and schemaType. Pure — exported for unit tests.
- */
-export function searchTemplates(
-  templates: ReadonlyArray<LauncherTemplate>,
-  query: string,
-): LauncherTemplate[] {
-  const trimmed = query.trim().toLowerCase()
-  if (!trimmed) return [...templates]
-  return templates.filter((template) => {
-    const haystack = `${template.title} ${template.ui.description} ${template.ui.group} ${template.schemaType}`.toLowerCase()
-    return haystack.includes(trimmed)
-  })
 }
 
 export function LauncherGrid({templates, onSelect}: LauncherGridProps): JSX.Element {
